jetstream-backend/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-aiplatform>=1.55.0
flask